Handle missing optional deps
