+++ tests/test_web_app.py
@@ -1,3 +1,6 @@
+import pytest
+
+fastapi = pytest.importorskip("fastapi")
 from fastapi.testclient import TestClient
 
 from web.main import app, index_path
