+++ web/main.py
@@ -10,6 +10,6 @@ app.mount("/static", StaticFiles(directory=static_dir), name="static")
 
 index_path = static_dir / "index.html"
 
-@app.get("/", response_class=HTMLResponse)
+@app.get("/", response_class=HTMLResponse)  # type: ignore[misc]
 async def index() -> str:
     return index_path.read_text(encoding="utf-8")
